requirements.txt
cycler==0.11.0
dataclasses==0.8
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.9.0.80
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.5.4
SimpleITK==2.1.1.2
six==1.16.0
torch==1.10.2
typing_extensions==4.1.1
